add code refactoring
