+++ package.json
@@ -10,7 +10,6 @@
     "@types/node": "^16.7.13",
     "mobx": "^6.9.0",
     "mobx-react": "^7.6.0",
-    "mobx-react-lite": "^3.4.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.10.0",
