+++ src/components/HowToStartPage/HowToStartPage.view.tsx
@@ -1,7 +1,8 @@
 import React from "react";
-import styles from './HowToStartPage.module.css';
+import styles from "./HowToStartPage.module.css";
 
 export const HowToStartPage = () => {
+	
 	return (
 		<div className={styles.container}>Здесь будет страница Как Начать</div>
 	);
