--- src/components/InfoPage/InfoPage.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import styles from './InfoPage.module.css';
-
-export const InfoPage = () => {
-	return (
-		<div className={styles.container}>Здесь будет страница справки</div>
-	);
-};
