+++ src/components/InfoPage/InfoPage.view.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import styles from "./InfoPage.module.css";
+
+export const InfoPage = () => {
+	
+	return <div className={styles.container}>Здесь будет страница справки</div>;
+};
