+++ src/helper/constants.ts
@@ -1,0 +1,5 @@
+export const DEFAULT_ISLAND_MAP_MATRIX = [
+  ["Water", "i023", "i022", "Water", "Water"],
+  ["Water", "Water", "Boat", "Water", "i022"],
+  ["i022", "Water", "Water", "i023", "Water"],
+];
